Feature/init ordering service

+++ src/Services/Basket/Basket.API/Infrastructure/Filters/GlobalExceptionFilter.cs
@@ -5,10 +5,6 @@ using Microsoft.AspNetCore.Mvc.Filters;
 using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
 using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Net;
-using System.Threading.Tasks;
 
 namespace Basket.API.Infrastructure.Filters
 {
@@ -27,7 +23,8 @@ namespace Basket.API.Infrastructure.Filters
             Exception exception = context.Exception;
             logger.LogError(new EventId(exception.HResult), exception, exception.Message);
 
-            if (exception is BasketDomainException) {
+            if (exception is BasketDomainException)
+            {
 
                 var problemDetails = new ValidationProblemDetails()
                 {
@@ -39,7 +36,7 @@ namespace Basket.API.Infrastructure.Filters
                 problemDetails.Errors.Add("DomainValidations", new string[] { context.Exception.Message.ToString() });
 
                 context.Result = new BadRequestObjectResult(problemDetails);
-                context.HttpContext.Response.StatusCode = (int)StatusCodes.Status400BadRequest;
+                context.HttpContext.Response.StatusCode = StatusCodes.Status400BadRequest;
             }
             else
             {
@@ -49,7 +46,7 @@ namespace Basket.API.Infrastructure.Filters
                     DeveloperMessage = env.IsDevelopment() ? context.Exception : default
                 };
                 context.Result = new BadRequestObjectResult(errorDetails);
-                context.HttpContext.Response.StatusCode = (int)StatusCodes.Status400BadRequest;
+                context.HttpContext.Response.StatusCode = StatusCodes.Status400BadRequest;
             }
 
             context.ExceptionHandled = true;
